LayoutTests/tiled-drawing/scrolling/scroll-snap/scroll-snap-mandatory-borders-expected.txt

Tests that the scroll-snap feature works properly in overflow regions.
 Testing scroll-snap glide for horizontalTarget:
PASS div scrolled to next window.
Testing scroll-snap snap for horizontalTarget:
PASS div honored snap points.
Testing scroll-snap glide for verticalTarget:
FAIL div did not honor snap points. Expected 300, but got 50
Testing scroll-snap snap for verticalTarget:
FAIL div did not snap back to proper location for verticalTarget. Expected 50, but got 0
PASS successfullyParsed is true

TEST COMPLETE
